+++ lib/easy-rest/api/action-result/action-result.d.ts
@@ -1,0 +1,5 @@
+import { Promise } from "es6-promise";
+import { ResponseMessage } from "./response-message";
+export interface IActionResult {
+    executeAsync(): Promise<ResponseMessage>;
+}