+++ lib/easy-rest/http/http-action-descriptor.d.ts
@@ -1,0 +1,6 @@
+import { IParameterBindingOptions } from "../decorators/binding/parameter-binding-options";
+export declare class HttpActionDescriptor {
+    actionName: string;
+    bindings: IParameterBindingOptions[];
+    returnType: any;
+}