+++ lib/easy-rest/util.d.ts
@@ -1,2 +1,2 @@
 export * from './util/path-builder';
-export * from './util/context-data-provider';
+export * from './util/http-context-provider';