+++ src/easy-rest/api/api-controller.ts
@@ -1,32 +1,29 @@
-import {Request} from "express";
 import {IActionResult} from "./action-result/action-result";
 import {OkResult} from "./action-result/ok-result";
 import {ErrorResult} from "./action-result/error-result";
 import {BadRequestResult} from "./action-result/bad-request-result";
 import {NotFoundResult} from "./action-result/not-found-result";
 import {IPrincipal} from "../security/principal/principal";
+import {HttpActionContext} from "../http/http-action-context";
+import {Request} from "express";
 
 export class ApiController {
-  private _req: Request;
-  private _user: IPrincipal;
+  _context: HttpActionContext;
+  // private _user: IPrincipal;
 
   constructor() {
   }
 
-  get requset(): Request {
-    return this._req;
-  }
-
-  set requset(value: Request) {
-    this._req = value;
+  get context(): HttpActionContext {
+    return this._context;
   }
 
   get user(): IPrincipal {
-    return this._user;
+    return this._context.httpContext.user;
   }
 
-  set user(value: IPrincipal) {
-    this._user = value;
+  get request(): Request {
+    return this._context.httpContext.request;
   }
 
   protected ok<T>(data?: any): IActionResult {