+++ src/easy-rest/http/http-controller-descriptor.ts
@@ -1,0 +1,6 @@
+import {IControllerConstructor} from "../api/api-controller";
+
+export class HttpControllerDescriptor {
+  controllerName: string;
+  controller: IControllerConstructor;
+}