+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './easy-rest/controller';
+export * from './easy-rest/api';
 export * from './easy-rest/core';
 export * from './easy-rest/decorators';
 export * from './easy-rest/handlers';